fix mod live wf

See merge request si/esup-signature!154

+++ src/main/java/org/esupportail/esupsignature/service/SignBookService.java
@@ -270,15 +270,22 @@ public class SignBookService {
         }
     }
 
-    public void removeStep(SignBook signBook, int step) {
-        LiveWorkflowStep liveWorkflowStep = signBook.getLiveWorkflow().getLiveWorkflowSteps().get(step);
-        signBook.getLiveWorkflow().getLiveWorkflowSteps().remove(liveWorkflowStep);
-        for(Recipient recipient : liveWorkflowStep.getRecipients()) {
-            for(SignRequest signRequest : signBook.getSignRequests()) {
-                signRequest.getRecipientHasSigned().remove(recipient);
+    @Transactional
+    public boolean removeStep(Long signBookId, int step) {
+        SignBook signBook = getById(signBookId);
+        int currentStepNumber = signBook.getLiveWorkflow().getCurrentStepNumber();
+        if(currentStepNumber <= step) {
+            LiveWorkflowStep liveWorkflowStep = signBook.getLiveWorkflow().getLiveWorkflowSteps().get(step);
+            signBook.getLiveWorkflow().getLiveWorkflowSteps().remove(liveWorkflowStep);
+            for (Recipient recipient : liveWorkflowStep.getRecipients()) {
+                for (SignRequest signRequest : signBook.getSignRequests()) {
+                    signRequest.getRecipientHasSigned().remove(recipient);
+                }
             }
+            liveWorkflowStepService.delete(liveWorkflowStep);
+            return true;
         }
-        liveWorkflowStepService.delete(liveWorkflowStep);
+        return false;
     }
 
     @Transactional
@@ -494,15 +501,15 @@ public class SignBookService {
     public void addLiveStep(Long id, String[] recipientsEmails, int stepNumber, Boolean allSignToComplete, String signType, boolean repeatable) throws EsupSignatureException {
         SignBook signBook = this.getById(id);
         int currentSetNumber = signBook.getLiveWorkflow().getCurrentStepNumber();
-        if(stepNumber > currentSetNumber) {
-            LiveWorkflowStep liveWorkflowStep = liveWorkflowStepService.createWorkflowStep(repeatable, allSignToComplete, SignType.valueOf(signType), recipientsEmails);
-            if (stepNumber == -1) {
-                signBook.getLiveWorkflow().getLiveWorkflowSteps().add(liveWorkflowStep);
-            } else {
-                signBook.getLiveWorkflow().getLiveWorkflowSteps().add(stepNumber - 1, liveWorkflowStep);
-            }
+        LiveWorkflowStep liveWorkflowStep = liveWorkflowStepService.createWorkflowStep(repeatable, allSignToComplete, SignType.valueOf(signType), recipientsEmails);
+        if (stepNumber == -1) {
+            signBook.getLiveWorkflow().getLiveWorkflowSteps().add(liveWorkflowStep);
         } else {
-            throw new EsupSignatureException("L'étape ne peut pas être ajoutée");
+            if (stepNumber >= currentSetNumber) {
+                signBook.getLiveWorkflow().getLiveWorkflowSteps().add(stepNumber - 1, liveWorkflowStep);
+            } else {
+                throw new EsupSignatureException("L'étape ne peut pas être ajoutée");
+            }
         }
     }
